move create_view content to create_app
