+++ datamodel/scripts/setup.sh
@@ -40,4 +40,4 @@ psql "service=${PGSERVICE}" -v ON_ERROR_STOP=1 -f ${DIR}/app/symbology_functions
 psql "service=${PGSERVICE}" -v ON_ERROR_STOP=1 -f ${DIR}/app/reach_direction_change.sql -v SRID=$SRID
 psql "service=${PGSERVICE}" -v ON_ERROR_STOP=1 -f ${DIR}/app/14_geometry_functions.sql -v SRID=$SRID
 
-${DIR}/app/create_tww_app.py --pg_service ${PGSERVICE} --srid ${SRID}
+${DIR}/app/create_app.py --pg_service ${PGSERVICE} --srid ${SRID}
